% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient.R
\name{agronomic_efficiency}
\alias{agronomic_efficiency}
\title{Agronomic Efficiency (AE)}
\usage{
agronomic_efficiency(
  yield_treated,
  yield_control,
  nutrient_applied,
  verbose = TRUE
)
}
\arguments{
\item{yield_treated}{Numeric vector. Yield with nutrient applied (kg/ha).}

\item{yield_control}{Numeric vector. Yield without nutrient, i.e., from
control plot (kg/ha).}

\item{nutrient_applied}{Numeric vector. Amount of nutrient applied (kg/ha).}

\item{verbose}{Logical. If \code{TRUE}, prints informational messages.
Default \code{TRUE}.}
}
\value{
Numeric vector. Agronomic efficiency (kg grain per kg nutrient).
}
\description{
Increase in economic yield per unit of nutrient applied.
}
\details{
\deqn{AE = \frac{Y_f - Y_0}{F}}{AE = (Yf - Y0) / F}
}
\examples{
agronomic_efficiency(yield_treated = 4500, yield_control = 3000,
                     nutrient_applied = 120)

}
\references{
Dobermann, A. (2007). Nutrient use efficiency: measurement and
management. In \emph{Fertilizer Best Management Practices}, IFA,
Paris, pp. 1-28.
\url{https://digitalcommons.unl.edu/agronomyfacpub/316/}

Fixen, P. et al. (2015). Nutrient/fertilizer use efficiency:
Measurement, current situation and trends. In \emph{Managing Water
and Fertilizer for Sustainable Agricultural Intensification},
IFA/IWMI/IPNI/IPI, pp. 8-38.

Congreves, K.A. et al. (2021). Nitrogen use efficiency definitions
of today and tomorrow. \emph{Frontiers in Plant Science}, 12,
637108. \doi{10.3389/fpls.2021.637108}
}
