% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.R
\name{carbon_sequestration_rate}
\alias{carbon_sequestration_rate}
\title{Carbon Sequestration Rate}
\usage{
carbon_sequestration_rate(soc_initial, soc_final, years, verbose = TRUE)
}
\arguments{
\item{soc_initial}{Numeric vector. Initial SOC stock (Mg C/ha).}

\item{soc_final}{Numeric vector. Final SOC stock (Mg C/ha).}

\item{years}{Numeric. Number of years between measurements.}

\item{verbose}{Logical. If \code{TRUE}, prints informational messages.
Default \code{TRUE}.}
}
\value{
Numeric vector. Carbon sequestration rate (Mg C/ha/year).
  Positive values indicate net sequestration; negative values
  indicate net emission from soil.
}
\description{
Annual rate of soil carbon change between two measurement time points.
}
\details{
Formula: \deqn{CSR = \frac{SOC_{final} - SOC_{initial}}{Years}}
}
\examples{
carbon_sequestration_rate(soc_initial = 28.5, soc_final = 31.2,
                          years = 5)

}
\references{
Minasny, B. et al. (2017). Soil carbon 4 per mille.
\emph{Geoderma}, 292, 59-86. \doi{10.1016/j.geoderma.2017.01.002}
}
