% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy_profitability}
\alias{energy_profitability}
\title{Energy Profitability}
\usage{
energy_profitability(gross_return, energy_in, verbose = TRUE)
}
\arguments{
\item{gross_return}{Numeric vector. Gross economic return (currency/ha).}

\item{energy_in}{Numeric vector. Total energy input (MJ/ha).}

\item{verbose}{Logical. If \code{TRUE}, prints informational messages.
Default \code{TRUE}.}
}
\value{
Numeric vector. Energy profitability (currency/MJ).
}
\description{
Economic return per unit of energy invested.
}
\details{
Formula: \deqn{EPr = \frac{Energy Output - Energy Input}{Energy Input}}{EPr = (Energy Output - Energy Input) / Energy Input}
}
\examples{
energy_profitability(gross_return = 112500, energy_in = 12000)

}
