% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy_use_efficiency}
\alias{energy_use_efficiency}
\title{Energy Use Efficiency (EUE)}
\usage{
energy_use_efficiency(energy_out, energy_in, verbose = TRUE)
}
\arguments{
\item{energy_out}{Numeric vector. Total energy output (MJ/ha).}

\item{energy_in}{Numeric vector. Total energy input (MJ/ha).}

\item{verbose}{Logical. If \code{TRUE}, prints informational messages.
Default \code{TRUE}.}
}
\value{
Numeric vector. Energy use efficiency (dimensionless ratio).
  Values greater than 1 indicate a positive energy balance.
}
\description{
Ratio of total energy output to total energy input.
}
\details{
\deqn{EUE = \frac{E_{out}}{E_{in}}}{EUE = E_out / E_in}
}
\examples{
energy_use_efficiency(energy_out = 135000, energy_in = 12000)

}
