% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.R
\name{land_equivalent_ratio}
\alias{land_equivalent_ratio}
\title{Land Equivalent Ratio (LER)}
\usage{
land_equivalent_ratio(
  yield_inter_a,
  yield_sole_a,
  yield_inter_b,
  yield_sole_b,
  verbose = TRUE
)
}
\arguments{
\item{yield_inter_a}{Numeric vector. Yield of crop A in intercrop
(kg/ha).}

\item{yield_sole_a}{Numeric vector. Yield of crop A in sole crop
(kg/ha).}

\item{yield_inter_b}{Numeric vector. Yield of crop B in intercrop
(kg/ha).}

\item{yield_sole_b}{Numeric vector. Yield of crop B in sole crop
(kg/ha).}

\item{verbose}{Logical. If \code{TRUE}, prints informational messages.
Default \code{TRUE}.}
}
\value{
Numeric vector. LER values. Values greater than 1 indicate
  an intercropping advantage; less than 1 indicates a disadvantage.
}
\description{
Evaluates the productivity advantage of intercropping systems
over sole cropping.
}
\details{
\deqn{LER = \frac{Y_{iA}}{Y_{sA}} + \frac{Y_{iB}}{Y_{sB}}}{LER = Y_iA/Y_sA + Y_iB/Y_sB}
}
\examples{
land_equivalent_ratio(yield_inter_a = 3500, yield_sole_a = 4500,
                      yield_inter_b = 800, yield_sole_b = 1200)

}
\references{
Mead, R. & Willey, R.W. (1980). The concept of a land equivalent
ratio and advantages in yields from intercropping. \emph{Experimental
Agriculture}, 16(3), 217-228. \doi{10.1017/S0014479700010978}
}
