% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{net_energy}
\alias{net_energy}
\title{Net Energy}
\usage{
net_energy(energy_out, energy_in, verbose = TRUE)
}
\arguments{
\item{energy_out}{Numeric vector. Total energy output (MJ/ha).}

\item{energy_in}{Numeric vector. Total energy input (MJ/ha).}

\item{verbose}{Logical. If \code{TRUE}, prints informational messages.
Default \code{TRUE}.}
}
\value{
Numeric vector. Net energy (MJ/ha). Positive values indicate
  net energy gain; negative values indicate net energy loss.
}
\description{
Difference between total energy output and total energy input.
}
\details{
Formula: \deqn{NE = Energy Output - Energy Input}{NE = Energy Output - Energy Input}
}
\examples{
net_energy(energy_out = 135000, energy_in = 12000)

}
