% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexus.R
\name{normalize_zscore}
\alias{normalize_zscore}
\title{Normalize Data (Z-score)}
\usage{
normalize_zscore(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of z-scores. Returns all zeros if standard
  deviation is zero.
}
\description{
Standardization to z-scores (mean = 0, standard deviation = 1).
}
\details{
Formula: \deqn{Z = \frac{X - \mu}{\sigma}}{Z = (X - mean) / sd}
}
\examples{
normalize_zscore(c(10, 20, 30, 40, 50))

}
