% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient.R
\name{nutrient_harvest_index}
\alias{nutrient_harvest_index}
\title{Nutrient Harvest Index (NHI)}
\usage{
nutrient_harvest_index(
  grain_nutrient_uptake,
  total_nutrient_uptake,
  verbose = TRUE
)
}
\arguments{
\item{grain_nutrient_uptake}{Numeric vector. Nutrient uptake in grain
(kg/ha).}

\item{total_nutrient_uptake}{Numeric vector. Total plant nutrient
uptake (kg/ha).}

\item{verbose}{Logical. If \code{TRUE}, prints informational messages.
Default \code{TRUE}.}
}
\value{
Numeric vector. Nutrient harvest index (proportion, 0 to 1).
}
\description{
Proportion of total plant nutrient partitioned to grain.
}
\details{
Formula: \deqn{HI = \frac{Economic Yield}{Biological Yield}}{HI = Economic Yield / Biological Yield}
}
\examples{
nutrient_harvest_index(grain_nutrient_uptake = 75,
                       total_nutrient_uptake = 100)

}
