% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.cf_result}
\alias{plot.cf_result}
\title{Plot method for cf_result}
\usage{
\method{plot}{cf_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cf_result}.}

\item{...}{Additional arguments passed to \code{\link[graphics]{barplot}}.}
}
\value{
No return value, called for the side effect of generating a
  horizontal bar plot showing carbon footprint by emission source.
}
\description{
Plot method for cf_result
}
