% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.nexus_result}
\alias{plot.nexus_result}
\title{Plot method for nexus_result}
\usage{
\method{plot}{nexus_result}(x, type = c("radar", "heatmap"), ...)
}
\arguments{
\item{x}{An object of class \code{nexus_result}.}

\item{type}{Character. Plot type: \code{"radar"} (default) or
\code{"heatmap"}.}

\item{...}{Additional arguments passed to \code{\link{nexus_radar}}
or \code{\link{nexus_heatmap}}.}
}
\value{
Invisibly returns the scores matrix. Called for the side
  effect of generating a radar or heatmap plot.
}
\description{
Plot method for nexus_result
}
