% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.cf_result}
\alias{print.cf_result}
\title{Print method for cf_result}
\usage{
\method{print}{cf_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cf_result}.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
Invisibly returns the input object \code{x}. Called for the
  side effect of printing a formatted carbon footprint summary to
  the console.
}
\description{
Print method for cf_result
}
