% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.nexus_result}
\alias{print.nexus_result}
\title{Print method for nexus_result}
\usage{
\method{print}{nexus_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{nexus_result}.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\value{
Invisibly returns the input object \code{x}. Called for the
  side effect of printing a formatted nexus analysis summary to the
  console.
}
\description{
Print method for nexus_result
}
