% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.nexus_result}
\alias{summary.nexus_result}
\title{Summary method for nexus_result}
\usage{
\method{summary}{nexus_result}(object, ...)
}
\arguments{
\item{object}{An object of class \code{nexus_result}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{object}. Called for
  the side effect of printing a detailed dimension-wise summary to
  the console.
}
\description{
Summary method for nexus_result
}
